console/requirements.txt
selenium==4.21.0
webdriver-manager==4.0.1
pillow==10.3.0